Simplify Installation using git.
