--- custom_components/hacs/frontend/__init__.py
@@ -1,1 +1,0 @@
-"""Initialize frontend objects."""
