--- custom_components/hacs/frontend/elements/__init__.py
@@ -1,1 +1,0 @@
-"""Initialize frontend elements."""
