--- custom_components/hacs/frontend/views/__init__.py
@@ -1,9 +1,0 @@
-"""Initialize frontend views."""
-from .error import HacsErrorView
-from .static import HacsStaticView
-from .plugin import HacsPluginView
-from .overview import HacsOverviewView
-from .store import HacsStoreView
-from .settings import HacsSettingsView
-from .repository import HacsRepositoryView
-from .api import HacsAPIView
