--- custom_components/hacs/handler/__init__.py
@@ -1,1 +1,0 @@
-"""Initialize handlers."""
