--- custom_components/hacs/manifest.json
@@ -1,8 +1,0 @@
-{
-  "domain": "hacs",
-  "name": "HACS (Home Assistant Community Store)",
-  "documentation": "https://custom-components.github.io/hacs",
-  "dependencies": [],
-  "codeowners": ["@ludeeus"],
-  "requirements": ["aiofiles", "backoff"]
-}
