+++ frontend/elements/__init__.py
@@ -1,0 +1,1 @@
+"""Initialize frontend elements."""
