+++ frontend/views/static.py
