+++ frontend/views/store.py
